+++ android/src/main/java/com/jimmydaddy/imagemarker/MarkerImageLoader.kt
@@ -5,77 +5,78 @@ import android.content.res.Resources
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.os.Build
+import android.os.Build.VERSION.SDK_INT
 import android.util.Log
 import androidx.annotation.RequiresApi
-import com.facebook.common.references.CloseableReference
-import com.facebook.datasource.DataSource
-import com.facebook.drawee.backends.pipeline.Fresco
-import com.facebook.imagepipeline.common.ResizeOptions
-import com.facebook.imagepipeline.core.ImagePipelineConfig
-import com.facebook.imagepipeline.datasource.BaseBitmapDataSubscriber
-import com.facebook.imagepipeline.image.CloseableImage
-import com.facebook.imagepipeline.request.ImageRequest
-import com.facebook.imagepipeline.request.ImageRequestBuilder
+import androidx.core.graphics.drawable.toBitmap
+import coil.ImageLoader
+import coil.decode.GifDecoder
+import coil.decode.ImageDecoderDecoder
+import coil.decode.SvgDecoder
+import coil.request.ImageRequest
 import com.facebook.react.bridge.ReactApplicationContext
-import com.facebook.react.modules.systeminfo.ReactNativeVersion
 import com.jimmydaddy.imagemarker.base.Constants.IMAGE_MARKER_TAG
 import com.jimmydaddy.imagemarker.base.ErrorCode
 import com.jimmydaddy.imagemarker.base.ImageOptions
 import com.jimmydaddy.imagemarker.base.MarkerError
-import com.jimmydaddy.imagemarker.base.Utils
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.withContext
 import java.util.concurrent.CompletableFuture
-import java.util.concurrent.Executor
-import java.util.concurrent.Executors
 
-class ImageLoader(private val context: ReactApplicationContext, private val maxSize: Int) {
+class MarkerImageLoader(private val context: ReactApplicationContext, private val maxSize: Int) {
 
-  init {
-    if (maxSize > 0) {
-      setMaxBitmapSize(maxSize)
+  private var imageLoader: ImageLoader = ImageLoader.Builder(context)
+    .components {
+      if (SDK_INT >= 28) {
+        add(ImageDecoderDecoder.Factory())
+      } else {
+        add(GifDecoder.Factory())
+      }
+      add(SvgDecoder.Factory())
     }
-  }
+    .allowHardware(false)
+    .build()
   private val resources: Resources
     get() = context.resources
 
   @RequiresApi(Build.VERSION_CODES.N)
   suspend fun loadImages(images: List<ImageOptions>): List<Bitmap?> = withContext(Dispatchers.IO) {
+
     val deferredList = images.map { img ->
       async {
         try {
-          val isFrescoImg = isFrescoImg(img.uri)
-          Log.d(IMAGE_MARKER_TAG, "isFrescoImg: " + isFrescoImg(img.uri))
-          if (isFrescoImg) {
+          val isCoilImg = isCoilImg(img.uri)
+          Log.d(IMAGE_MARKER_TAG, "isCoilImg: $isCoilImg")
+          if (isCoilImg) {
             val future = CompletableFuture<Bitmap?>()
-            var imageRequest = ImageRequest.fromUri(img.uri)
+            var request = ImageRequest.Builder(context)
+              .data(img.uri)
             if (img.src != null && img.src.width > 0 && img.src.height > 0) {
-              val options: ResizeOptions? = ResizeOptions(img.src.width, img.src.height)
-              imageRequest = ImageRequestBuilder.fromRequest(imageRequest).setResizeOptions(options).build()
+              request = request.size(img.src.width, img.src.height)
               Log.d(IMAGE_MARKER_TAG, "src.width: " + img.src.width + " src.height: " + img.src.height)
             }
-            val dataSource = Fresco.getImagePipeline().fetchDecodedImage(imageRequest, null)
-            val executor: Executor = Executors.newSingleThreadExecutor()
-            dataSource.subscribe(object : BaseBitmapDataSubscriber() {
-              public override fun onNewResultImpl(bitmap: Bitmap?) {
-                if (bitmap != null) {
-                  val bg = ImageProcess.scaleBitmap(bitmap, img.scale)
-                  future.complete(bg)
-                } else {
+            imageLoader.enqueue(request.target (
+              onStart = { _ ->
+                // Handle the placeholder drawable.
+                Log.d(IMAGE_MARKER_TAG, "start to load image: " + img.uri)
+              },
+              onSuccess = { result ->
+                val bitmap = result.toBitmap()
+                val bg = ImageProcess.scaleBitmap(bitmap, img.scale)
+                if (bg == null) {
                   future.completeExceptionally(MarkerError(ErrorCode.LOAD_IMAGE_FAILED,
-                    "Can't retrieve the file from the src: " + img.uri))
+                      "Can't retrieve the file from the src: " + img.uri))
                 }
-              }
-
-              override fun onFailureImpl(dataSource: DataSource<CloseableReference<CloseableImage>>) {
+                future.complete(bg)
+              },
+              onError = { _ ->
                 future.completeExceptionally(MarkerError(ErrorCode.LOAD_IMAGE_FAILED,
                   "Can't retrieve the file from the src: " + img.uri))
               }
-            }, executor)
+            ).build())
             return@async future.get()
-
           } else {
             val resId = getDrawableResourceByName(img.uri)
             Log.d(IMAGE_MARKER_TAG, "resId: $resId")
@@ -109,7 +110,7 @@ class ImageLoader(private val context: ReactApplicationContext, private val maxS
     deferredList.awaitAll()
   }
 
-  private fun isFrescoImg(uri: String?): Boolean {
+  private fun isCoilImg(uri: String?): Boolean {
     // val base64Pattern =
     // "^data:(image|img)/(bmp|jpg|png|tif|gif|pcx|tga|exif|fpx|svg|psd|cdr|pcd|dxf|ufo|eps|ai|raw|WMF|webp);base64,(([[A-Za-z0-9+/])*\\s\\S*)*"
     return uri!!.startsWith("http://") || uri.startsWith("https://") || uri.startsWith("file://") || uri.startsWith(
@@ -124,21 +125,6 @@ class ImageLoader(private val context: ReactApplicationContext, private val maxS
       "drawable",
       context.packageName
     )
-  }
-
-  private fun setMaxBitmapSize(maxSize: Int) {
-    val major = Utils.getStringSafe("major", ReactNativeVersion.VERSION)
-    val minor = Utils.getStringSafe("minor", ReactNativeVersion.VERSION)
-    val patch = Utils.getStringSafe("patch", ReactNativeVersion.VERSION)
-    if (Integer.valueOf(major.toString()) >= 0 && Integer.valueOf(minor.toString()) >= 60 && Integer.valueOf(
-        patch.toString()
-      ) >= 0
-    ) {
-      val config =
-        ImagePipelineConfig.newBuilder(context).experiment().setMaxBitmapSize(maxSize)
-          .build()
-      Fresco.initialize(context, config)
-    }
   }
 
 }